% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_methods.r
\name{run_adroit}
\alias{run_adroit}
\title{Run AdRoit deconvolution method}
\usage{
run_adroit(scrna_path, spatial_path, celltype_final, output_path)
}
\arguments{
\item{scrna_path}{Path to scRNA-seq data}

\item{spatial_path}{Path to spatial transcriptomics data}

\item{celltype_final}{Column name in sc_obj@meta.data containing cell type information}

\item{output_path}{Path to save output}
}
\value{
AdRoit result
}
\description{
Run AdRoit deconvolution method
}
\examples{
library(AdRoit)

scrna_path = './sc_rna.h5seurat'
spatial_path = './st_data.h5seurat'
celltype_final = 'celltype'
output_path = 'output/'
sc_obj <- LoadH5Seurat(scrna_path)
spatial_obj <- LoadH5Seurat(spatial_path)
sc_count <- as.matrix(sc_obj@assays$RNA@counts)
spatial_count <- as.matrix(spatial_obj@assays$RNA@counts)
spatial_location <- data.frame(colnames(spatial_obj))
spatial_location <- data.frame(spatial_obj@meta.data)

cellType <-  sc_obj@meta.data[,celltype_final]
    
sc_count <- as.matrix(sc_count)
spatial_count <- as.matrix(spatial_count)
    
single.ref = ref.build(sc_count,
                   cellType,
                   genes=intersect(rownames(spatial_count),rownames(sc_count)),
                   normalize = "None",
                   multi.sample.bulk = TRUE,
                   multi.sample.single = TRUE,
                   nbootsids=5, 
                   minbootsize=50,
                   silent = FALSE)

print("single cell ref done")
st_count_ma=as.matrix(spatial_count)
st_count_ma = st_count_ma[intersect(rownames(st_count_ma),rownames(sc_count)),]

registerDoSEQ()
result=AdRoit.est(
     st_count_ma,
     single.ref,
     use.refvar = FALSE,
     per.sample.adapt = FALSE,
     silent = TRUE)

}