% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{run_deconvolution_methods}
\alias{run_deconvolution_methods}
\title{Run all deconvolution methods}
\usage{
run_deconvolution_methods(
  scrna_path,
  spatial_path,
  celltype_final,
  python_path,
  output_path = tempdir()
)
}
\arguments{
\item{scrna_path}{Path to scRNA-seq data}

\item{spatial_path}{Path to spatial transcriptomics data}

\item{celltype_final}{Column name in sc_obj@meta.data containing cell type information}

\item{python_path}{Path to Python executable}

\item{output_path}{Directory to save output, If not set, save to the temporary directory tempdir().}
}
\value{
all deconvolution result
}
\description{
Run all deconvolution methods
}
