% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_methods.r
\name{run_redeconve}
\alias{run_redeconve}
\title{Run Redeconve deconvolution method}
\usage{
run_redeconve(scrna_path, spatial_path, celltype_final, output_path)
}
\arguments{
\item{scrna_path}{Path to scRNA-seq data}

\item{spatial_path}{Path to spatial transcriptomics data}

\item{celltype_final}{Column name in sc_obj@meta.data containing cell type information}

\item{output_path}{Path to save output}
}
\value{
Redeconve result
}
\description{
Run Redeconve deconvolution method
}
\examples{
scrna_path = './sc_rna.h5seurat'
spatial_path = './st_data.h5seurat'
celltype_final = 'celltype'
output_path = 'output/'

sc_obj <- LoadH5Seurat(scrna_path)
spatial_obj <- LoadH5Seurat(spatial_path)
sc_count <- as.matrix(sc_obj@assays$RNA@counts)
spatial_count <- as.matrix(spatial_obj@assays$RNA@counts)
spatial_location <- data.frame(colnames(spatial_obj))
spatial_location <- data.frame(spatial_obj@meta.data)
    
colnames(spatial_location) <- c('x', 'y')
cellType <-  sc_obj@meta.data
cellType$barcodes <- rownames(cellType)
cellType <- cellType[, c("barcodes", colnames(cellType)[-ncol(cellType)])]
    
## get reference
ref <- get.ref(sc_count, cellType, dopar = FALSE)
  
    
## deconvolution
res.ct <- deconvoluting(ref, spatial_count, genemode = "def", hpmode = "auto", dopar = TRUE, ncores = 8)
res.prop <- to.proportion(res.ct)
Redeconve_result <- t(res.prop)
save(Redeconve_result, file = file.path(output_path, "Redeconve_result.Rdata"))
}
