% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_methods.r
\name{run_seurat}
\alias{run_seurat}
\title{Run Seurat deconvolution method}
\usage{
run_seurat(scrna_path, spatial_path, celltype_final, output_path)
}
\arguments{
\item{scrna_path}{Path to scRNA-seq data}

\item{spatial_path}{Path to spatial transcriptomics data}

\item{celltype_final}{Column name in sc_obj@meta.data containing cell type information}

\item{output_path}{Path to save output}
}
\value{
Seurat result
}
\description{
Run Seurat deconvolution method
}
\examples{
scrna_path = './sc_rna.h5seurat'
spatial_path = './st_data.h5seurat'
celltype_final = 'celltype'
output_path = 'output/'
sc_rna <- LoadH5Seurat(scrna_path)
spatial <- LoadH5Seurat(spatial_path)
    
sc_rna <- SCTransform(sc_rna)
spatial <- SCTransform(spatial)
anchors <- FindTransferAnchors(reference = sc_rna, 
                               query = spatial, 
                               k.anchor = 10, 
                               k.filter = 200)
  
predictions <- TransferData(anchorset = anchors, 
                            refdata = sc_rna@meta.data[,celltype_final], 
                            dims = 1:30)
Seurat_result <- predictions[,which(colnames(predictions) != "predicted.id" & colnames(predictions) != "prediction.score.max")]
colnames(Seurat_result) = gsub("prediction.score.", "", colnames(Seurat_result))
save(Seurat_result, file = file.path(output_path,"Seurat_result.Rdata"))
}
