% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{run_deconvolution_methods}
\alias{run_deconvolution_methods}
\title{Run all deconvolution methods}
\usage{
run_deconvolution_methods(
  scrna_path,
  spatial_path,
  celltype_final,
  python_path,
  output_path
)
}
\arguments{
\item{scrna_path}{Path to scRNA-seq data}

\item{spatial_path}{Path to spatial transcriptomics data}

\item{celltype_final}{Column name in sc_obj@meta.data containing cell type information}

\item{python_path}{Path to Python executable}

\item{output_path}{Directory to save output}
}
\value{
all deconvolution result
}
\description{
Run all deconvolution methods
}
\examples{
scrna_path = './sc_rna.h5seurat'
spatial_path = './st_data.h5seurat'
celltype_final = 'celltype'
ptrhon_path = "D:/Application/anaconda3/envs/python_3.10"
output_path = 'output/'

# Run all methods
run_deconvolution_methods(scrna_path, spatial_path,celltype_final, python_path, output_path)
}
