% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_methods.r
\name{run_rctd}
\alias{run_rctd}
\title{Run RCTD deconvolution method}
\usage{
run_rctd(scrna_path, spatial_path, celltype_final, output_path)
}
\arguments{
\item{scrna_path}{Path to scRNA-seq data}

\item{spatial_path}{Path to spatial transcriptomics data}

\item{celltype_final}{Column name in sc_obj@meta.data containing cell type information}

\item{output_path}{Path to save output}
}
\value{
RCTD result
}
\description{
Run RCTD deconvolution method
}
