% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_methods.r
\name{run_spatialdwls}
\alias{run_spatialdwls}
\title{Run SpatialDWLS deconvolution method}
\usage{
run_spatialdwls(
  scrna_path,
  spatial_path,
  celltype_final,
  python_path,
  output_path
)
}
\arguments{
\item{scrna_path}{Path to scRNA-seq data}

\item{spatial_path}{Path to spatial transcriptomics data}

\item{celltype_final}{Column name in sc_obj@meta.data containing cell type information}

\item{python_path}{Path to Python executable}

\item{output_path}{Path to save output}
}
\value{
SpatialDWLS result
}
\description{
Run SpatialDWLS deconvolution method
}
